+++ inst/tinytest/test-ZM.R
@@ -1,7 +1,4 @@
-context("conversion")
-
-
-test_that("sfc objects with ZM converted to GeoJSON", {
+## "sfc objects with ZM converted to GeoJSON", {
 
 	js <- '{"type":"Point","coordinates":[0.0,0.0]}'
 	sf <- geojson_sfc(js)
@@ -146,11 +143,9 @@ test_that("sfc objects with ZM converted to GeoJSON", {
 	expect_equal( attr( sf[[1]], "class" )[1], "XYZM")
 	j <- sfc_geojson( sf )
 	expect_true( jsonify::validate_json( j ) )
-})
 
 
-## SF ZM
-test_that("sf objects with ZM converted to GeoJSON", {
+## "sf objects with ZM converted to GeoJSON", {
 
 	js <- '{"type":"Point","coordinates":[0.0,0.0]}'
 	sf <- geojson_sf(js)
@@ -295,11 +290,9 @@ test_that("sf objects with ZM converted to GeoJSON", {
 	expect_equal( attr( sf$geometry[[1]], "class" )[1], "XYZM")
 	j <- sf_geojson( sf )
 	expect_true( jsonify::validate_json( j ) )
-})
-## SF ZM
 
 
-test_that("geometry collections with geometries with XYZM dimensions", {
+## "geometry collections with geometries with XYZM dimensions", {
 
 	js <- '{"type":"GeometryCollection","geometries":[{"type":"MultiPolygon","coordinates":[[[[0,0],[0,1],[1,1],[1,0],[0,0,0]]]]}]}'
 	sf <- geojson_sfc( js )
@@ -317,10 +310,7 @@ test_that("geometry collections with geometries with XYZM dimensions", {
 	expect_equal( attr(sf[[1]][[1]],  "class")[1], "XYZM") ## MULTIPOLYGON has XYZM
 	expect_equal( attr(sf[[1]][[2]],  "class")[1], "XY") ## MULTIPOLYGON has XY
 
-})
-
-
-test_that("ZM attributes", {
+## "ZM attributes", {
 
 	pt <- '{"type":"Point","coordinates":[0.1,0.1]}'
 	expect_true( length( as.numeric( geojson_sfc( pt )[[1]] ) ) == 2 )
@@ -361,43 +351,52 @@ test_that("ZM attributes", {
 
 	mls <- '{"type":"MultiLineString","coordinates":[[[0.0,0],[1,1]]]}'
 	expect_equal( as.matrix( geojson_sfc( mls )[[1]][[1]] ), matrix(c(0,0,1,1), ncol = 2, byrow = T) )
+
 	mls <- '{"type":"MultiLineString","coordinates":[[[0.0,0,1],[1,1]]]}'
-	expect_equal( as.matrix( geojson_sfc( mls )[[1]][[1]] ), matrix(c(0,0,1,1,1,0), ncol = 3, byrow = T) )
+	expect_equal( as.matrix( geojson_sfc( mls )[[1]][[1]] ), matrix(c(0,0,1,1,1,NA), ncol = 3, byrow = T) )
+
 	mls <- '{"type":"MultiLineString","coordinates":[[[0.0,0],[1,1,2]]]}'
-	expect_equal( as.matrix( geojson_sfc( mls )[[1]][[1]] ), matrix(c(0,0,0,1,1,2), ncol = 3, byrow = T) )
+	expect_equal( as.matrix( geojson_sfc( mls )[[1]][[1]] ), matrix(c(0,0,NA,1,1,2), ncol = 3, byrow = T) )
+
 	mls <- '{"type":"MultiLineString","coordinates":[[[0.0,0,1],[1,1,2]]]}'
 	expect_equal( as.matrix( geojson_sfc( mls )[[1]][[1]] ), matrix(c(0,0,1,1,1,2), ncol = 3, byrow = T) )
+
 	mls <- '{"type":"MultiLineString","coordinates":[[[0.0,0,1,2],[1,1,2]]]}'
-	expect_equal( as.matrix( geojson_sfc( mls )[[1]][[1]] ), matrix(c(0,0,1,2,1,1,2,0), ncol = 4, byrow = T) )
+	expect_equal( as.matrix( geojson_sfc( mls )[[1]][[1]] ), matrix(c(0,0,1,2,1,1,2,NA), ncol = 4, byrow = T) )
+
 	mls <- '{"type":"MultiLineString","coordinates":[[[0.0,0],[1,1,2,3]]]}'
-	expect_equal( as.matrix( geojson_sfc( mls )[[1]][[1]] ), matrix(c(0,0,0,0,1,1,2,3), ncol = 4, byrow = T) )
+	expect_equal( as.matrix( geojson_sfc( mls )[[1]][[1]] ), matrix(c(0,0,NA,NA,1,1,2,3), ncol = 4, byrow = T) )
+
 	mls <- '{"type":"MultiLineString","coordinates":[[[0.0,0,1,2],[1,1,2,3]]]}'
 	expect_equal( as.matrix( geojson_sfc( mls )[[1]][[1]] ), matrix(c(0,0,1,2,1,1,2,3), ncol = 4, byrow = T) )
 
 	ply <- '{"type":"Polygon","coordinates":[[[0,0],[0,1],[1,1],[1,0],[0,0]]]}'
 	expect_equal( as.matrix( geojson_sfc( ply )[[1]][[1]] ), matrix(c(0,0,0,1,1,1,1,0,0,0), ncol = 2, byrow = T) )
+
 	ply <- '{"type":"Polygon","coordinates":[[[0,0,1],[0,1],[1,1],[1,0],[0,0]]]}'
-	expect_equal( as.matrix( geojson_sfc( ply )[[1]][[1]] ), matrix(c(0,0,1,0,1,0,1,1,0,1,0,0,0,0,0), ncol = 3, byrow = T) )
+	expect_equal( as.matrix( geojson_sfc( ply )[[1]][[1]] ), matrix(c(0,0,1,0,1,NA,1,1,NA,1,0,NA,0,0,NA), ncol = 3, byrow = T) )
+
 	ply <- '{"type":"Polygon","coordinates":[[[0,0,1],[0,1],[1,1],[1,0],[0,0,9]]]}'
-	expect_equal( as.matrix( geojson_sfc( ply )[[1]][[1]] ), matrix(c(0,0,1,0,1,0,1,1,0,1,0,0,0,0,9), ncol = 3, byrow = T) )
+	expect_equal( as.matrix( geojson_sfc( ply )[[1]][[1]] ), matrix(c(0,0,1,0,1,NA,1,1,NA,1,0,NA,0,0,9), ncol = 3, byrow = T) )
+
 	ply <- '{"type":"Polygon","coordinates":[[[0,0,1],[0,1],[1,1],[1,0],[0,0,9]]]}'
-	expect_equal( as.matrix( geojson_sfc( ply )[[1]][[1]] ), matrix(c(0,0,1,0,1,0,1,1,0,1,0,0,0,0,9), ncol = 3, byrow = T) )
+	expect_equal( as.matrix( geojson_sfc( ply )[[1]][[1]] ), matrix(c(0,0,1,0,1,NA,1,1,NA,1,0,NA,0,0,9), ncol = 3, byrow = T) )
+
 	ply <- '{"type":"Polygon","coordinates":[[[0,0],[0,1],[1,1,3],[1,0],[0,0]]]}'
-	expect_equal( as.matrix( geojson_sfc( ply )[[1]][[1]] ), matrix(c(0,0,0,0,1,0,1,1,3,1,0,0,0,0,0), ncol = 3, byrow = T) )
+	expect_equal( as.matrix( geojson_sfc( ply )[[1]][[1]] ), matrix(c(0,0,NA,0,1,NA,1,1,3,1,0,NA,0,0,NA), ncol = 3, byrow = T) )
+
 	ply <- '{"type":"Polygon","coordinates":[[[0,0],[0,1],[1,1,3],[1,0,1,2],[0,0]]]}'
-	expect_equal( as.matrix( geojson_sfc( ply )[[1]][[1]] ), matrix(c(0,0,0,0,0,1,0,0,1,1,3,0,1,0,1,2,0,0,0,0), ncol = 4, byrow = T) )
+	expect_equal( as.matrix( geojson_sfc( ply )[[1]][[1]] ), matrix(c(0,0,NA,NA,0,1,NA,NA,1,1,3,NA,1,0,1,2,0,0,NA,NA), ncol = 4, byrow = T) )
 
 	mply <- '{"type":"MultiPolygon","coordinates":[[[[0,0],[0,1],[1,1],[1,0],[0,0]],[[2,2],[2,3,9],[3,3],[3,2],[2,2]]]]}'
 	expect_equal( geojson_sfc( mply )[[1]][[1]][[1]]  , matrix(c(0,0,0,1,1,1,1,0,0,0), ncol = 2, byrow = T) )
-	expect_equal( geojson_sfc( mply )[[1]][[1]][[2]]  , matrix(c(2,2,0,2,3,9,3,3,0,3,2,0,2,2,0), ncol = 3, byrow = T) )
+	expect_equal( geojson_sfc( mply )[[1]][[1]][[2]]  , matrix(c(2,2,NA,2,3,9,3,3,NA,3,2,NA,2,2,NA), ncol = 3, byrow = T) )
 
 	mply <- '{"type":"MultiPolygon","coordinates":[[[[0,0],[0,1],[1,1,2,2],[1,0],[0,0]],[[2,2],[2,3,9],[3,3],[3,2],[2,2]]]]}'
-	expect_equal( geojson_sfc( mply )[[1]][[1]][[1]]  , matrix(c(0,0,0,0,0,1,0,0,1,1,2,2,1,0,0,0,0,0,0,0), ncol = 4, byrow = T) )
-	expect_equal( geojson_sfc( mply )[[1]][[1]][[2]]  , matrix(c(2,2,0,2,3,9,3,3,0,3,2,0,2,2,0), ncol = 3, byrow = T) )
+	expect_equal( geojson_sfc( mply )[[1]][[1]][[1]]  , matrix(c(0,0,NA,NA,0,1,NA,NA,1,1,2,2,1,0,NA,NA,0,0,NA,NA), ncol = 4, byrow = T) )
+	expect_equal( geojson_sfc( mply )[[1]][[1]][[2]]  , matrix(c(2,2,NA,2,3,9,3,3,NA,3,2,NA,2,2,NA), ncol = 3, byrow = T) )
 
-})
-
-test_that("NA values handled", {
+##"NA values handled", {
 
 
 	sf <- structure(list(geometry = structure(list(structure(c(0, 1, 0,
@@ -412,7 +411,49 @@ test_that("NA values handled", {
 	expect_true( jsonify::validate_json( geo ) )
 	expect_equal( as.character( geo ), '{"type":"LineString","coordinates":[[0.0,0.0,0.0,0.0],[1.0,null,null,1.0]]}' )
 
-})
+
+## Issue 85
+## z & m NA when not initialised
+
+	zm <- function(x) {
+		all( is.na( unclass( unname(x) ) ) )
+	}
+	geojson <- '{ "type":"Point","coordinates":[0,0] }'
+	z <- attr( geojson_sfc( geojson ), "z_range")
+	m <- attr( geojson_sfc( geojson ), "m_range")
+	expect_true( zm( z ) )
+	expect_true( zm( m ) )
+
+	geojson <- '{ "type":"Point","coordinates":[0,0,0] }'
+	z <- attr( geojson_sfc( geojson ), "z_range")
+	m <- attr( geojson_sfc( geojson ), "m_range")
+	expect_false( zm( z ) )
+	expect_true( zm( m ) )
+
+	geojson <- '{ "type":"Point","coordinates":[0,0,0,0] }'
+	z <- attr( geojson_sfc( geojson ), "z_range")
+	m <- attr( geojson_sfc( geojson ), "m_range")
+	expect_false( zm( z ) )
+	expect_false( zm( m ) )
+
+
+	geojson <- '{ "type":"LineString","coordinates":[[0,0],[1,1]] }'
+	z <- attr( geojson_sfc( geojson ), "z_range")
+	m <- attr( geojson_sfc( geojson ), "m_range")
+	expect_true( zm( z ) )
+	expect_true( zm( m ) )
+
+	geojson <- '{ "type":"LineString","coordinates":[[0,0,0],[1,1,1]] }'
+	z <- attr( geojson_sfc( geojson ), "z_range")
+	m <- attr( geojson_sfc( geojson ), "m_range")
+	expect_false( zm( z ) )
+	expect_true( zm( m ) )
+
+	geojson <- '{ "type":"LineString","coordinates":[[0,0,0,0],[1,1,1,1]] }'
+	z <- attr( geojson_sfc( geojson ), "z_range")
+	m <- attr( geojson_sfc( geojson ), "m_range")
+	expect_false( zm( z ) )
+	expect_false( zm( m ) )